change map file structure
