+++ src/pages/admin/admin-dontion-centers/add-donation-cennter-dialog/donation-center-map/map.tsx
@@ -1,0 +1,69 @@
+import { isLatLngLiteral } from "@googlemaps/typescript-guards";
+import { Children, cloneElement, isValidElement, useEffect, useRef, useState } from "react";
+import { createCustomEqual } from "fast-equals";
+
+interface MapProps extends google.maps.MapOptions {
+  className?: string;
+  style: { [key: string]: string };
+  onClick?: (e: google.maps.MapMouseEvent) => void;
+  onIdle?: (map: google.maps.Map) => void;
+}
+
+export const Map: React.FC<MapProps> = ({ onClick, onIdle, children, className, style, ...options }) => {
+  const ref = useRef<HTMLDivElement>(null);
+  const [map, setMap] = useState<google.maps.Map>();
+
+  useEffect(() => {
+    if (ref.current && !map) {
+      setMap(new window.google.maps.Map(ref.current, {}));
+    }
+  }, [ref, map]);
+
+  useDeepCompareEffectForMaps(() => {
+    if (map) {
+      map.setOptions(options);
+    }
+  }, [map, options]);
+
+  useEffect(() => {
+    if (map) {
+      ["click", "idle"].forEach((eventName) => google.maps.event.clearListeners(map, eventName));
+      if (onClick) {
+        map.addListener("click", onClick);
+      }
+      if (onIdle) {
+        map.addListener("idle", () => onIdle(map));
+      }
+    }
+  }, [map, onClick, onIdle]);
+
+  return (
+    <div className={className}>
+      <div ref={ref} style={style} />
+      {Children.map(children, (child) => {
+        if (isValidElement(child)) {
+          return cloneElement(child, { map });
+        }
+      })}
+    </div>
+  );
+};
+
+const deepCompareEqualsForMaps = createCustomEqual((deepEqual) => (a: any, b: any) => {
+  if (isLatLngLiteral(a) || a instanceof google.maps.LatLng || isLatLngLiteral(b) || b instanceof google.maps.LatLng) {
+    return new google.maps.LatLng(a).equals(new google.maps.LatLng(b));
+  }
+  return deepEqual(a, b);
+});
+
+function useDeepCompareMemoize(value: any) {
+  const ref = useRef();
+  if (!deepCompareEqualsForMaps(value, ref.current)) {
+    ref.current = value;
+  }
+  return ref.current;
+}
+
+function useDeepCompareEffectForMaps(callback: React.EffectCallback, dependencies: any[]) {
+  useEffect(callback, dependencies.map(useDeepCompareMemoize));
+}
