+++ src/pages/admin/admin-dontion-centers/add-donation-cennter-dialog/donation-center-map/marker.tsx
@@ -1,0 +1,24 @@
+import { useEffect, useState } from "react";
+
+export const Marker: React.FC<google.maps.MarkerOptions> = (options) => {
+  const [marker, setMarker] = useState<google.maps.Marker>();
+
+  useEffect(() => {
+    if (!marker) {
+      setMarker(new google.maps.Marker());
+    }
+    return () => {
+      if (marker) {
+        marker.setMap(null);
+      }
+    };
+  }, [marker]);
+
+  useEffect(() => {
+    if (marker) {
+      marker.setOptions(options);
+    }
+  }, [marker, options]);
+
+  return null;
+};
